+++ widgets/TaskRoleInfoBox.php
@@ -5,6 +5,7 @@ namespace humhub\modules\tasks\widgets;
 
 
 use humhub\modules\tasks\models\Task;
+use humhub\modules\ui\icon\widgets\Icon;
 use Yii;
 
 class TaskRoleInfoBox extends TaskInfoBox
@@ -24,13 +25,13 @@ class TaskRoleInfoBox extends TaskInfoBox
     public function getValue()
     {
          if ( $this->task->isTaskResponsible()) {
-             return '<i class="fa fa-check"></i> '.Yii::t('TasksModule.base', 'You are responsible!');
+             return Icon::get('check')->color($this->view->theme->variable('success')).' '.Yii::t('TasksModule.base', 'You are responsible!');
          } else if($this->task->isTaskAssigned()) {
-             return '<i class="fa fa-check"></i> '.Yii::t('TasksModule.base', 'You are assigned!');
+             return Icon::get('check')->color($this->view->theme->variable('success')).' '.Yii::t('TasksModule.base', 'You are assigned!');
          } else if($this->task->canProcess()) {
-            return  '<i class="fa fa-times"></i> '.Yii::t('TasksModule.widgets_views_wallentry', 'Anyone can work on this task!');
+            return  Icon::get('times')->color($this->view->theme->variable('danger')).' '.Yii::t('TasksModule.widgets_views_wallentry', 'Anyone can work on this task!');
          } else {
-             return  '<i class="fa fa-times"></i> '.Yii::t('TasksModule.widgets_views_wallentry', 'This task can only be processed by assigned and responsible users.');
+             return  Icon::get('times')->color($this->view->theme->variable('danger')).' '.Yii::t('TasksModule.widgets_views_wallentry', 'This task can only be processed by assigned and responsible users.');
          }
     }